// src/app/types/weather-gov/weatherGov.d.ts
export interface paths {
// ...
  '/points/{point}': {
    parameters: {
      query?: never;
      header?: never;
      path: {
        /** @description Point (latitude, longitude) */
        point: components['parameters']['PathPoint'];
      };
      cookie?: never;
    };
    /** @description Returns metadata about a given latitude/longitude point */
    get: operations['point'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
}